// src/api/model/eventGeneric.ts
/**
 * Generated by orval v6.25.0 🍺
 * Do not edit manually.
 * Ticketerie API
 * OpenAPI spec version: 1.0.0
 */
import type { Location } from './location';
import type { Sport } from './sport';
import type { Ticket } from './ticket';

export interface EventGeneric {
  createdAt: string;
  eventDate: string;
  icon: string;
  id: string;
  location: Location;
  locationId: number;
  name: string;
  sport: Sport;
  sportId: number;
  tickets: Ticket[];
  updatedAt: string | null;
}