// src/api/model/eventPlayer.ts
/**
 * Generated by orval v6.25.0 🍺
 * Do not edit manually.
 * Ticketerie API
 * OpenAPI spec version: 1.0.0
 */
import type { Championship } from './championship';
import type { Location } from './location';
import type { Player } from './player';
import type { Sport } from './sport';
import type { Ticket } from './ticket';

export interface EventPlayer {
  championship: Championship;
  championshipId: number;
  createdAt: string;
  eventDate: string;
  id: string;
  location: Location;
  locationId: number;
  player1: Player;
  player2: Player;
  sport: Sport;
  sportId: number;
  tickets: Ticket[];
  updatedAt: string | null;
}