// src/api/model/eventTeam.ts
/**
 * Generated by orval v6.25.0 🍺
 * Do not edit manually.
 * Ticketerie API
 * OpenAPI spec version: 1.0.0
 */
import type { Team } from './team';
import type { Championship } from './championship';
import type { Location } from './location';
import type { Sport } from './sport';
import type { Ticket } from './ticket';

export interface EventTeam {
  awayTeam: Team;
  awayTeamId: number;
  championship: Championship;
  championshipId: number;
  createdAt: string;
  eventDate: string;
  homeTeam: Team;
  homeTeamId: number;
  id: string;
  location: Location;
  locationId: number;
  sport: Sport;
  sportId: number;
  tickets: Ticket[];
  updatedAt: string | null;
}