// src/api/model/ticket.ts
/**
 * Generated by orval v6.25.0 🍺
 * Do not edit manually.
 * Ticketerie API
 * OpenAPI spec version: 1.0.0
 */
import type { Ticketing } from './ticketing';

export interface Ticket {
  createdAt: string;
  eventId: string;
  id: number;
  price: number;
  ticketing: Ticketing;
  ticketingId: number;
  updatedAt?: string | null;
  url: string;
}